+++ ChangeLog.md
@@ -1,3 +1,6 @@
 # Changelog for turtle-service
 
 ## Unreleased changes
+
+- 17.04.20: Added TimeSeries GET Endpoint
+- 15.04.20: Initial Setup: GET/POST Status Endpoints